Implement Windows NodePublish/Unpublish
